lib: fwts_framework: use size_t instead of int
Acked-by: Alex Hung <alex.hung@canonical.com>
Acked-by: Keng-Yu Lin <kengyu@canonical.com>

+++ src/lib/src/fwts_framework.c
@@ -786,7 +786,7 @@ static void fwts_framework_heading_info(fwts_framework *fw, fwts_list *tests_to_
 	time_t now;
 	struct utsname buf;
 	char *tests = NULL;
-	int len = 1;
+	size_t len = 1;
 	int i;
 	fwts_list_link *item;
